fixed test setup
