+++ core/test/com/group66/game/cannon/ColoredBallTest.java
@@ -4,12 +4,36 @@ import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 import static org.mockito.Mockito.mock;
 
+import org.junit.Before;
 import org.junit.Test;
 
+import com.badlogic.gdx.Gdx;
+import com.badlogic.gdx.backends.headless.HeadlessFiles;
+import com.badlogic.gdx.backends.headless.HeadlessNativesLoader;
+import com.badlogic.gdx.backends.headless.HeadlessNet;
+import com.badlogic.gdx.backends.headless.mock.graphics.MockGraphics;
+import com.badlogic.gdx.graphics.GL20;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 import com.group66.game.BustaMove;
 
 public class ColoredBallTest extends BallTest {
+    @Before
+    public void setupTests() {
+        HeadlessNativesLoader.load();
+        MockGraphics mockGraphics = new MockGraphics();
+        Gdx.graphics = mockGraphics;
+        HeadlessNet headlessNet = new HeadlessNet();
+        Gdx.net = headlessNet;
+        HeadlessFiles headlessFiles = new HeadlessFiles();
+        Gdx.files = headlessFiles;
+        Gdx.gl = mock(GL20.class);
+        try {
+            BustaMove.getGameInstance().create();
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+    }
+    
     @Test
     public void creationTest() {
         new ColoredBall(BallType.RED, 0, 0, 0, 0);
