+++ core/test/com/group66/game/screencontrollers/AbstractGameControllerTest.java
@@ -1,20 +1,43 @@
 package com.group66.game.screencontrollers;
 
+import static org.mockito.Mockito.mock;
+
+import org.junit.Before;
 import org.junit.Test;
 
+import com.badlogic.gdx.Gdx;
+import com.badlogic.gdx.backends.headless.HeadlessFiles;
+import com.badlogic.gdx.backends.headless.HeadlessNativesLoader;
+import com.badlogic.gdx.backends.headless.HeadlessNet;
+import com.badlogic.gdx.backends.headless.mock.graphics.MockGraphics;
+import com.badlogic.gdx.graphics.GL20;
 import com.group66.game.BustaMove;
 import com.group66.game.cannon.GameManager;
 
 public abstract class AbstractGameControllerTest {
-    protected abstract AbstractGameController getGameController();
     
-    @Test
-    public void pauseTest() {
+    @Before
+    public void setupTests() {
+        HeadlessNativesLoader.load();
+        MockGraphics mockGraphics = new MockGraphics();
+        Gdx.graphics = mockGraphics;
+        HeadlessNet headlessNet = new HeadlessNet();
+        Gdx.net = headlessNet;
+        HeadlessFiles headlessFiles = new HeadlessFiles();
+        Gdx.files = headlessFiles;
+        Gdx.gl = mock(GL20.class);
         try {
             BustaMove.getGameInstance().create();
         } catch (Exception e) {
             e.printStackTrace();
         }
+    }
+    
+    protected abstract AbstractGameController getGameController();
+    
+    @Test
+    public void pauseTest() {
+        
         AbstractGameController controller = getGameController();
         controller.togglePause();
         controller.update(0);
@@ -22,22 +45,14 @@ public abstract class AbstractGameControllerTest {
     
     @Test
     public void updateTest() {
-        try {
-            BustaMove.getGameInstance().create();
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
+        
         AbstractGameController controller = getGameController();
         controller.update(0);
     }
     
     @Test
     public void winTest() {
-        try {
-            BustaMove.getGameInstance().create();
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
+        
         AbstractGameController controller = getGameController();
         try {
             GameManager gameManager = controller.getGameManager1();
@@ -50,11 +65,7 @@ public abstract class AbstractGameControllerTest {
     
     @Test
     public void loseTest() {
-        try {
-            BustaMove.getGameInstance().create();
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
+        
         AbstractGameController controller = getGameController();
         try {
             GameManager gameManager = controller.getGameManager1();
