+++ core/test/com/group66/game/screencontrollers/TwoPlayerGameControllerTest.java
@@ -1,9 +1,9 @@
 package com.group66.game.screencontrollers;
 
 import org.junit.Test;
-/*
-public class TwoPlayerGameControllerTest extends AbstractGameControllerTest {
 
+public class TwoPlayerGameControllerTest extends AbstractGameControllerTest {
+    
     @Override
     protected AbstractGameController getGameController() {
         return new TwoPlayerGameController(false, 3);
@@ -14,4 +14,3 @@ public class TwoPlayerGameControllerTest extends AbstractGameControllerTest {
         new TwoPlayerGameController(true, 1);
     }
 }
-*/
